% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_probs.R
\name{add_probs}
\alias{add_probs}
\title{Function to predict prababilities for each sampled data set
Take a model and a predicted set of training data and calculate the probabilities of fishing 
in each site given a set choice set.}
\usage{
add_probs(train_dat, the_model, ncores = 6)
}
\arguments{
\item{train_dat}{Training data, sampled from training_data function}

\item{the_model}{Output from sampled_rums function}

\item{ncores}{Number of cores to run in parallel, defaults to 6}
}
\description{
Function to predict prababilities for each sampled data set
Take a model and a predicted set of training data and calculate the probabilities of fishing 
in each site given a set choice set.
}
