% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_probs.R
\name{adjust_probs}
\alias{adjust_probs}
\title{Adjust Cluster Probabilities
Function to adjust cluster probabilities based on different things}
\usage{
adjust_probs(rum1, quotas1, rc1)
}
\arguments{
\item{rum1}{RUM results}

\item{quotas1}{Quota amounts}

\item{rc1}{risk_coefficient
risk coefficients;
-1 -- blind fishing
0-- only use original data values
1--take catches into account; will eventually get towards 10
10 -- avoid risk fishing}
}
\description{
Adjust Cluster Probabilities
Function to adjust cluster probabilities based on different things
}
