% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_list.R
\name{arg_list}
\alias{arg_list}
\title{Argument List
Function to generate list of arguments for bigger function}
\usage{
arg_list(m_y = 2007, f_y = 2009, nhauls_sampled = 50, ncores = 8,
  seed = the_seed, r_c = 50, r_s = 100, ports = the_ports,
  h_d = the_hd, dyz = the_days, quota_species = c("Canary Rockfish"),
  n_c = "trev")
}
\arguments{
\item{m_y}{minimum year}
}
\description{
Argument List
Function to generate list of arguments for bigger function
}
