% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bycatch.R
\name{calc_bycatch}
\alias{calc_bycatch}
\title{Calculate Bycatch
Function to calculate bycatch from sampled tows}
\usage{
calc_bycatch(nreps = 5000, nsamps = 50, tac = 50000, seed = 300,
  of_interest)
}
\arguments{
\item{nreps}{Number of replicates}

\item{nsamps}{Number of samples in each replicate}

\item{tac}{Total allowable catch value}

\item{seed}{Seed value}

\item{of_interest}{Data frame of interest, filtered by target species or location possibly}
}
\description{
Calculate Bycatch
Function to calculate bycatch from sampled tows
}
\examples{
Put example in
}

