% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_comb_deltas.R
\name{calc_comb_deltas}
\alias{calc_comb_deltas}
\title{Calculate Delta values for each vessel combination in each cluster
Function to calculate Delta values for each vessel combination in each cluster}
\usage{
calc_comb_deltas(ll)
}
\arguments{
\item{ll}{is a row index in the data frame vess_combs, 
which has combinations of each vessel in each cluster}
}
\description{
Calculate Delta values for each vessel combination in each cluster
Function to calculate Delta values for each vessel combination in each cluster
}
