% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ctac_rev.R
\name{calc_ctac_rev}
\alias{calc_ctac_rev}
\title{Function to calculate catch-quota balancing and revenues
Filters the data so that the catches remain below the tac. Catches are allowed to go 
over the tac for one species.}
\usage{
calc_ctac_rev(quotas, catches)
}
\arguments{
\item{quotas}{Data frame of the quota amounts for each species}

\item{catches}{Resampled catches from RUM}
}
\description{
Function to calculate catch-quota balancing and revenues
Filters the data so that the catches remain below the tac. Catches are allowed to go 
over the tac for one species.
}
