% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_probs.R
\name{calc_probs}
\alias{calc_probs}
\title{Calculate Probabilities 
Calculate probabilites based on type of profits and type of catch}
\usage{
calc_probs(prof_type = "avg_profit_fuel_only", catch_type,
  objective = "difference", in_cp = poss_clusts)
}
\arguments{
\item{prof_type}{Type of profit; can be "avg_haul_profit" or "avg_profit_fuel_only"}

\item{catch_type}{Type of catch; can be "type_clust_catch", "type_clust_perc", or "type_prop_hauls"}

\item{objective}{Objective; to catch the "highest" of target species? "lowest" of weak stock species? 
Or area with the biggest "difference" between target and weaks?}

\item{in_cp}{Input to calc_probs; defaults to poss_clusts}
}
\description{
Calculate Probabilities 
Calculate probabilites based on type of profits and type of catch
}
\examples{
f1 <- find_best_clusts(catch_type = "type_clust_perc", objective = "highest")
f2 <- find_best_clusts(catch_type = "type_prop_hauls", objective = "highest")
f3 <- find_best_clusts(catch_type = "type_clust_perc", objective = "lowest")
f4 <- find_best_clusts(catch_type = "type_prop_hauls", objective = "lowest")
#Results might be slightly different for different columns
cbind(f3$unq_clust, f4$unq_clust)
}
