% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4_perm_test.R
\name{ch4_perm_test}
\alias{ch4_perm_test}
\title{Permutation Test
Function to conduct permutation test on different columns}
\usage{
ch4_perm_test(input, column, ndraws = 1000)
}
\arguments{
\item{input}{Input data frame}

\item{column}{Column to permute}

\item{ndraws}{Number of draws; defaults to 1000}
}
\description{
Permutation Test
Function to conduct permutation test on different columns
}
\examples{
dd <- ch4_perm_test(input = top100_clusts, column = 'ntows', ndraws = 50)
}
