% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_two.R
\name{change_two}
\alias{change_two}
\title{Function to change two things at once
Function to change two things at once}
\usage{
change_two(in_list_start, to_change1, to_change2, values1, values2)
}
\arguments{
\item{in_list_start}{Starting list of arguments}

\item{to_change1}{Thing one to change}

\item{to_change2}{Thing two to change}

\item{values1}{Values for to_change1}

\item{values2}{Values for to_change2}
}
\description{
Function to change two things at once
Function to change two things at once
}
