% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_by_port.R
\name{clust_by_port}
\alias{clust_by_port}
\title{Cluster values by port
Function to cluster values by port. Can specify a cut_point or use a default calculation
from Trevor's fishing opportunity paper}
\usage{
clust_by_port(port, cut_point = NA)
}
\arguments{
\item{port}{Name or index for port of interest}

\item{cut_point}{Default is calculation based on median tow distance}
}
\description{
Cluster values by port
Function to cluster values by port. Can specify a cut_point or use a default calculation
from Trevor's fishing opportunity paper
}
