% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_scope.R
\name{clust_scope}
\alias{clust_scope}
\title{Define scope of potential new clusters
Function to define scope of potential new clusters}
\usage{
clust_scope(catch_input, clust_scope, input_cs)
}
\arguments{
\item{catch_input}{Catch input data, from catch_list}

\item{clust_scope}{Scope of possible new clusters}

\item{input_cs}{Input data used to deetermine next clusters}
}
\description{
Define scope of potential new clusters
Function to define scope of potential new clusters
}
