% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_aic.R
\name{compare_aic}
\alias{compare_aic}
\title{Function to compare AIC and LL values}
\usage{
compare_aic(risk_coefficient, net_cost, seed, nhauls, years = 2009:2014,
  nports = 4, dir = "/Volumes/udrive/", ncores = 6)
}
\arguments{
\item{risk_coefficient}{Risk coefficient}

\item{net_cost}{Net cost adjustment}

\item{seed}{Seed}

\item{nhauls}{Number of hauls}

\item{years}{Years to evaluate}

\item{nports}{Number of ports}

\item{dir}{Path to directory}

\item{ncores}{Number of cores to use}
}
\description{
Function to compare AIC and LL values
}
