% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_fleet.R
\name{fish_fleet}
\alias{fish_fleet}
\title{Function to resample averages for each vessel in a fleet}
\usage{
fish_fleet(fleet_chars = vess_vals, rum_res = the_probs)
}
\arguments{
\item{fleet_chars}{Vessel charcteristics, can filter here if fewere samplees for each
vessel desired vess_vals}

\item{rum_res}{RUM results, formatted as list}
}
\description{
Function to resample averages for each vessel in a fleet
}
