% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_trip_simulation.R
\name{fish_trip_simulation}
\alias{fish_trip_simulation}
\title{Run Fish Trip Simulation
Function to run fish_trip simulations in parallel}
\usage{
fish_trip_simulation(in_list)
}
\arguments{
\item{nreps}{Number of replicates; each nrep corresponds to a seed}

\item{ncores}{Number of cores; used in mclapply}

\item{ntows}{Number of tows; start of input into fish_trip function}

\item{scope}{Scope of movement}

\item{input}{Overall input to the function; defaults to filt_clusts which is the data with clusters filtered to include clusters with the most tows}

\item{ntows}{Number of tows in the fishing trip}

\item{start_clust}{Starting cluster to fish in}

\item{quotas}{Data frame of quota for target and weak stock species}

\item{scale}{Scale of movement; if scale == 'port', specify a d_port; if scale == 'scope',  specify scale}

\item{prob_type}{Probability type; type_prop_hauls -- frequency of encounter 
or type_clust_perc -- proportion of catch in each cluster}
}
\description{
Run Fish Trip Simulation
Function to run fish_trip simulations in parallel
}
\examples{
#Specify arguments as a list
in_list <- list(nreps = 6, ncores = 6, ntows = 20, scope = 5, 
  scale = 'scope', the_port = "ASTORIA / WARRENTON", prob_type = "type_clust_perc",
  quotas = quotas)
run_fish_trip_simulation(in_list)
}
