% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rum.R
\name{fit_rum}
\alias{fit_rum}
\title{Fit Random Utility model
Fit Random Utility Model to data. Uses multinomial logistic regression in package "nnet"}
\usage{
fit_rum(haul_dat, max_iter = 10000, max_weights = 3000,
  print_trace = FALSE)
}
\arguments{
\item{haul_dat}{Data to be input}

\item{max_iter}{Max number of iterations}

\item{max_weights}{Max. number of weights}

\item{print_trace}{If TRUE, print tracke}
}
\description{
Fit Random Utility model
Fit Random Utility Model to data. Uses multinomial logistic regression in package "nnet"
}
