% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_rum_data.R
\name{format_rum_data}
\alias{format_rum_data}
\title{Function to format RUM data
Function calls mlogit}
\usage{
format_rum_data(data_in = filt_clusts, the_port = "ASTORIA / WARRENTON",
  min_year = 2011, max_year = 2012, risk_coefficient = 1,
  tow_num_range = c(1, 1), ndays = 60)
}
\arguments{
\item{data_in}{Data going in to the function; default is filt_clusts}

\item{the_port}{Port of focus; Default is Astoria}

\item{min_year}{Minimum year used to filter the data}

\item{max_year}{Maximum year used to filter the data, also RUM data is filtered to be from the max_year}

\item{risk_coefficient}{Coefficient to adjust the quota prices up or down, feeding into net revenue calculations}

\item{tow_num_range}{Filter the tows to be the first tow or subsequent tows}

\item{ndays}{Number of previous days data to use in revenue expectations}

\item{trip_dists}{Distances covered by each trip}
}
\description{
Function to format RUM data
Function calls mlogit
}
