% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aic.R
\name{get_aic}
\alias{get_aic}
\title{Function to calculate predictions and AIC values
input a port value}
\usage{
get_aic(the_port, the_seed = 300)
}
\arguments{
\item{the_port}{Port name}

\item{the_seed}{Seed value for rmultinom and predictions}
}
\description{
Function to calculate predictions and AIC values
input a port value
}
