% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_choice_set.R, R/get_choicie_set.R
\name{get_choice_set}
\alias{get_choice_set}
\alias{get_choice_set}
\title{Generate Choice Set
Generate Choice Set for Predictions}
\usage{
get_choice_set(data_in = filt_clusts, the_port = "ASTORIA / WARRENTON",
  min_year = 2010, max_year = 2012, risk_coefficient = 1, ndays = 60,
  focus_year = 2012, nhauls_sampled = 50, seed = 300, ncores, rev_scale,
  model_type = "no_bycatch", net_cost, habit_distance)

get_choice_set(data_in = filt_clusts, the_port = "ASTORIA / WARRENTON",
  min_year = 2010, max_year = 2012, risk_coefficient = 1, ndays = 60,
  focus_year = 2012, nhauls_sampled = 50, seed = 300, ncores, rev_scale,
  model_type = "no_bycatch", net_cost, habit_distance)
}
\arguments{
\item{data_in}{Data going in to the function; default is filt_clusts}

\item{the_port}{Port of focus; Default is Astoria}

\item{min_year}{Minimum year used to filter the data}

\item{max_year}{Maximum year used to filter the data, also RUM data is filtered to be from the max_year}

\item{risk_coefficient}{Coefficient to adjust the quota prices up or down, feeding into net revenue calculations}

\item{ndays}{Number of previous days data to use in revenue expectations}

\item{focus_year}{Year to focus on for the models}

\item{nhauls_sampled}{Number of hauls to sample from the full data set}

\item{seed}{Seed for sampling tows}

\item{ncores}{Number of cores to use}

\item{rev_scale}{Scale the revenue by this factor}

\item{net_cost}{Type of netting of costs;}

\item{habit_distance}{Distance of spatiotemporal filter}

\item{trip_dists}{Distances covered by each trip}

\item{data_in}{Data going in to the function; default is filt_clusts}

\item{trip_dists}{Distances covered by each trip}

\item{the_port}{Port of focus; Default is Astoria}

\item{min_year}{Minimum year used to filter the data}

\item{max_year}{Maximum year used to filter the data, also RUM data is filtered to be from the max_year}

\item{risk_coefficient}{Coefficient to adjust the quota prices up or down, feeding into net revenue calculations}

\item{ndays}{Number of previous days data to use in revenue expectations}

\item{focus_year}{Year to focus on for the models}

\item{nhauls_sampled}{Number of hauls to sample from the full data set}

\item{seed}{Seed for sampling tows}

\item{ncores}{Number of cores to use}

\item{rev_scale}{Scale the revenue by this factor}

\item{net_cost}{Type of netting of costs;}

\item{habit_distance}{Distance of spatiotemporal filter}
}
\description{
Generate Choice Set
Generate Choice Set for Predictions

Generate Choice Set
Generate Choice Set for Predictions
}
