% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filenames.R
\name{get_filenames}
\alias{get_filenames}
\title{Function to get filenames to read in model results}
\usage{
get_filenames(the_args, years = 2011:2014, ports = NA)
}
\arguments{
\item{the_args}{List of the arguments for particular runs}

\item{years}{Vector of years; default to 2011:2014}
}
\description{
Function to get filenames to read in model results
}
