% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_df.R
\name{list_to_df}
\alias{list_to_df}
\title{Function to convert a list to dataframe
ldply but gives option to name columns different things}
\usage{
list_to_df(the_thing, ind_name, col_ind_name)
}
\arguments{
\item{the_thing}{Input}

\item{ind_name}{Index value names; like 'rep' or 'tow'}

\item{col_ind_name}{column name of index values}
}
\description{
Function to convert a list to dataframe
ldply but gives option to name columns different things
}
