% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_coefs.R
\name{load_coefs}
\alias{load_coefs}
\title{Load coefficients function
Function to load coefficients from model runs}
\usage{
load_coefs(the_args = a_l, res_dir = "/Volumes/udrive/")
}
\arguments{
\item{the_args}{List of arguments}

\item{res_dir}{Results directory}
}
\description{
Load coefficients function
Function to load coefficients from model runs
}
