% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clusts.R
\name{plot_clusts}
\alias{plot_clusts}
\title{Plot Clusters
Function to plot clusters}
\usage{
plot_clusts(dat, show_fig = TRUE, add_map = TRUE, specify_lims = TRUE)
}
\arguments{
\item{dat}{Input}

\item{show_fig}{If TRUE, show figure in device}

\item{add_map}{If TRUE, add clusters to map}

\item{specify_lims}{If TRUE, specify x and y limits from the data}
}
\description{
Plot Clusters
Function to plot clusters
}
