% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Function to plot pairs of values
Function to plot pairs of values}
\usage{
plot_pairs(var_pp = "spp_clust_catch", save_pp = TRUE)
}
\arguments{
\item{var_pp}{Variable of interest; Could by spp_clust_cathc}

\item{save_pp}{Save plot as png or not}
}
\description{
Function to plot pairs of values
Function to plot pairs of values
}
