% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paths.R
\name{plot_paths}
\alias{plot_paths}
\title{Plot fished clusters
Function to plot the paths of vessels. Currently configured to run with fish_trip_simulation}
\usage{
plot_paths(results_in, facet = FALSE, remove_legend = TRUE,
  add_map = FALSE, ylims = c(43, 49), xlims = c(-126, -124))
}
\arguments{
\item{facet}{Facet the figure?}

\item{remove_legend}{Remove the legend if TRUE}

\item{add_map}{Add map to the figure?}

\item{ylims}{y axis limits for map}

\item{xlims}{x axis limits for map}

\item{result_in}{Results of fish_trip_simulation}
}
\description{
Plot fished clusters
Function to plot the paths of vessels. Currently configured to run with fish_trip_simulation
}
