% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profits.R
\name{plot_profits}
\alias{plot_profits}
\title{Plot Profits
Function to plot profits associated with replicates. Currently configured to run with fish_trip}
\usage{
plot_profits(results_in, things)
}
\arguments{
\item{result_in}{Results of fish_trip_simulation}
}
\description{
Plot Profits
Function to plot profits associated with replicates. Currently configured to run with fish_trip
}
