% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_rums.R
\name{port_rums}
\alias{port_rums}
\title{port_rums function
Function to run and save RUM data for 8 top ports}
\usage{
port_rums(m_y, f_y, nhauls_sampled = 75, seed, ncores, r_c = 1, r_s = 10,
  ports, dum_type = "no_bycatch", dyz, h_d)
}
\arguments{
\item{m_y}{Minimum year}

\item{f_y}{Focus Year, also used as the maximum year}

\item{nhauls_sampled}{Number of hauls to sample}

\item{seed}{value to set seed}

\item{ncores}{Defaults to 6}

\item{r_c}{Risk Coefficient, defaults to 1}

\item{r_s}{Revenue scaling value, defaults to 10}

\item{dyz}{Days}

\item{h_d}{habit distance}
}
\description{
port_rums function
Function to run and save RUM data for 8 top ports
}
