% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_metrics.R
\name{pred_metrics}
\alias{pred_metrics}
\title{Prediction Metrics
Function that calculates metrics based on the tow with the highest modeled probability}
\usage{
pred_metrics(choices, model)
}
\arguments{
\item{choices}{Choice set}

\item{model}{Model results}
}
\description{
Prediction Metrics
Function that calculates metrics based on the tow with the highest modeled probability
}
