% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aic.R
\name{process_aic}
\alias{process_aic}
\title{Function to process AIC and LL values}
\usage{
process_aic(filename, dir = "/Volumes/udrive/")
}
\arguments{
\item{filename}{Name of file}

\item{dir}{Path to directory}
}
\description{
Function to process AIC and LL values
}
