% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coefficients.R
\name{process_coefficients}
\alias{process_coefficients}
\title{Function to process coefficients}
\usage{
process_coefficients(filename, dir = "/Volumes/udrive/")
}
\arguments{
\item{filename}{Name of file}

\item{dir}{Path to directory}
}
\description{
Function to process coefficients and add in relevant arguments
}
