% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dummys.R
\name{process_dummys}
\alias{process_dummys}
\title{Process dummy variables
Function to calculate revenues and distances}
\usage{
process_dummys(xx, td2 = td1, dat1 = dat)
}
\arguments{
\item{xx}{Index value}

\item{td2}{The tow dates}
}
\description{
Process dummy variables
Function to calculate revenues and distances
}
