% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files_predict.R
\name{read_files_predict}
\alias{read_files_predict}
\title{Read files and calculate model predictions
Function to read in model results and generate figures}
\usage{
read_files_predict(the_args, data_in = tows_clust, return_filenames = FALSE)
}
\arguments{
\item{the_args}{Argument List}

\item{data_in}{Data in for sampling}

\item{return_filenames}{Option to only return the filenames and not sample}
}
\description{
Read files and calculate model predictions
Function to read in model results and generate figures
}
