% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_training_data.R
\name{resample_training_data}
\alias{resample_training_data}
\title{Function to resample tows given a training data set 
Take models and training data set. Resample to see what average revenues/ catch:TAC are.}
\usage{
resample_training_data(training_data = dat_set1, rum_results = rc1,
  seed = 300, quotas = quotas_mb, port = "NEWPORT", rc)
}
\arguments{
\item{training_data}{Set of training data}

\item{rum_results}{Results from one of the RUM models}

\item{seed}{Seed for running replicates}
}
\description{
Function to resample tows given a training data set 
Take models and training data set. Resample to see what average revenues/ catch:TAC are.
}
