% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rum_probs.R
\name{rum_probs}
\alias{rum_probs}
\title{rum_probs Function
Function to format rum data, fit models, then return the probabilities of fishing in each place}
\usage{
rum_probs(rc = 1, port, years, fc = filt_clusts, ndays1 = 60)
}
\arguments{
\item{rc}{Risk Coefficient, used to scale costs of quota pounds}

\item{port}{Port of interest}

\item{years}{Vector of years, min_year and max_year}

\item{fc}{filt_clusts is default data in}

\item{ndays1}{Number of days to look back, default to 60}
}
\description{
rum_probs Function
Function to format rum data, fit models, then return the probabilities of fishing in each place
}
