% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_six_years.R
\name{run_six_years}
\alias{run_six_years}
\title{Run Six Years Function
Run six years of model with different conditions}
\usage{
run_six_years(the_args = a_l)
}
\arguments{
\item{the_args}{List of arguments created by arg_list}
}
\description{
Run Six Years Function
Run six years of model with different conditions
}
