% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_catches.R
\name{sample_catches}
\alias{sample_catches}
\title{Function to run the catch resampling
Order of functions is: 1) Fit RUM, 2) Calculate number of trips and tows for each vessel
3) Resample catches based on the probabilities in parallel}
\usage{
sample_catches(ctl_list, the_dat = filt_clusts)
}
\arguments{
\item{ctl_list}{List of arguments generated by function ch4_ctl}

\item{the_dat}{Data input, defaults to filt_clusts}
}
\description{
Function to run the catch resampling
Order of functions is: 1) Fit RUM, 2) Calculate number of trips and tows for each vessel
3) Resample catches based on the probabilities in parallel
}
