% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_hauls.R
\name{sample_hauls}
\alias{sample_hauls}
\title{Function to Sample Hauls
Function to Sample Hauls}
\usage{
sample_hauls(xx, hauls1 = hauls,
  dist_hauls_catch_shares1 = dist_hauls_catch_shares, nhauls_sampled1 = 50,
  depth_bin_proportions)
}
\arguments{
\item{xx}{Index}

\item{hauls1}{hauls data frame}

\item{dist_hauls_catch_shares1}{df with distance after catch shares}

\item{nhauls_sampled1}{number of sampled hauls}

\item{depth_bin_proportions}{Proportions of tows in each depth bin}
}
\description{
Function to Sample Hauls
Function to Sample Hauls
}
