% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_tows.R
\name{sample_tows}
\alias{sample_tows}
\title{Sample Tows
Function to sample tows from logbook data}
\usage{
sample_tows(nreps = 5000, nsamps = 50, tac = 50000, seed = 300)
}
\arguments{
\item{nreps}{Number of replicates}

\item{nsamps}{Number of samples in each replicate}

\item{tac}{Total allowable catch value}
}
\description{
Sample Tows
Function to sample tows from logbook data
}
\examples{
sample_tows(5, 10, 1000)
}
