% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampled_rums_clust.R
\name{sampled_rums_clust}
\alias{sampled_rums_clust}
\title{Format RUM data base on resampled tows
Function calls mlogit}
\usage{
sampled_rums_clust(data_in = filt_clusts, the_port = "ASTORIA / WARRENTON",
  min_year = 2011, max_year = 2012, risk_coefficient = 1, ndays = 60,
  focus_year = 2012, nhauls_sampled = 50, seed = 300, ncores)
}
\arguments{
\item{data_in}{Data going in to the function; default is filt_clusts}

\item{the_port}{Port of focus; Default is Astoria}

\item{min_year}{Minimum year used to filter the data}

\item{max_year}{Maximum year used to filter the data, also RUM data is filtered to be from the max_year}

\item{risk_coefficient}{Coefficient to adjust the quota prices up or down, feeding into net revenue calculations}

\item{ndays}{Number of previous days data to use in revenue expectations}

\item{focus_year}{Year to focus on for the models}

\item{nhauls_sampled}{Number of hauls to sample from the full data set}

\item{seed}{Seed for sampling tows}

\item{trip_dists}{Distances covered by each trip}
}
\description{
Format RUM data base on resampled tows
Function calls mlogit
}
