% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_permute.R
\name{simple_permute}
\alias{simple_permute}
\title{Function for simple permutations}
\usage{
simple_permute(input, perm_column, nperms = 1000, seed = 12345,
  crit = "<=", column = "set_year", index = c(2007, 2010))
}
\arguments{
\item{input}{The data to permute}

\item{perm_column}{Column for permutation}

\item{nperms}{Number of permutations/resamples}

\item{seed}{Seed for resampling}

\item{crit}{Criterion for evaluating significance}

\item{column}{Column of years by default, whatever index to permute}

\item{index}{Values in column that are indices of permutation}
}
\description{
Simple being that the data have to be in a specific format
}
