% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_corrs.R
\name{spp_corrs}
\alias{spp_corrs}
\title{Calculate Bycatch
Function to calculate catch correlations between two species}
\usage{
spp_corrs(tows = tows, output, two_species = c("dover sole", "sablefish"))
}
\arguments{
\item{tows}{Data frame of tows with species and catches}

\item{output}{Output from calc_bycatch function}

\item{two_species}{Two species to calculate correlations}
}
\description{
Calculate Bycatch
Function to calculate catch correlations between two species
}
\examples{
Put example in
}

