% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_delts.R
\name{spp_delts}
\alias{spp_delts}
\title{Function to calculate deltas for species
spp_delts}
\usage{
spp_delts(sppz, managed2)
}
\arguments{
\item{sppz}{Species}

\item{managed2}{Managed data frame}
}
\description{
Function to calculate deltas for species
spp_delts
}
