% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_for_probs.R
\name{values_for_probs}
\alias{values_for_probs}
\title{Values for probability calculations
Function to format values for calculating movement probabilities}
\usage{
values_for_probs(poss_clusts, input_vfb)
}
\arguments{
\item{poss_clust}{Vector of possible clusters}

\item{thing_vfb}{Input (filt_clusts essentailly)}
}
\description{
Values for probability calculations
Function to format values for calculating movement probabilities
}
