% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whn_spp_delta.R, R/year_spp_delta.R
\name{year_spp_delta}
\alias{year_spp_delta}
\alias{year_spp_delta}
\title{Function to calculate values for delta plots
Skew and proportion of zeroes specifically}
\usage{
year_spp_delta(year, spp)

year_spp_delta(year, spp)
}
\arguments{
\item{year}{Year to focus on}

\item{when}{Time period to focus on}

\item{skew}{Skew of non-zero, logged pound values}

\item{skew}{Skew of non-zero, logged pound values}
}
\description{
Function to calculate values for delta plots
Skew and proportion of zeroes specifically

Function to calculate values for delta plots
Skew and proportion of zeroes specifically
}
