% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4_perm_test.R
\name{ch4_perm_test}
\alias{ch4_perm_test}
\title{Permutation Test
Function to conduct permutation test on different columns}
\usage{
ch4_perm_test(input, column, ndraws = 1000, gb = "set_year, unq_clust",
  summ = "length(species)", clust_cat = "unq_clust", seed = 12345,
  crit = "<=", annual = TRUE, save_resamps = FALSE)
}
\arguments{
\item{input}{Input data frame}

\item{column}{Column to permute}

\item{ndraws}{Number of draws; defaults to 1000}

\item{gb}{Character string of things to group by for the parse statements}

\item{summ}{How to group things for resampling? Either by cluster, or cluster and type of species.}

\item{seed}{Random sampling seed}

\item{annual}{Resample annual averages (TRUE)? Or resample tows then calculate annual averages (FALSE)}

\item{save_resamps}{Save resampled values?}

\item{cri}{Criteria for determing p-value, should be "<" or "<="}
}
\description{
Permutation Test
Function to conduct permutation test on different columns
}
\examples{
dd <- ch4_perm_test(input = top100_clusts, column = 'ntows', ndraws = 50)
}
