% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_trip.R
\name{fish_trip}
\alias{fish_trip}
\title{Function to conduct a fishing trip
Function to conduct a fishing trip}
\usage{
fish_trip(input = filt_clusts, ntows = 10, start_clust = 295,
  seed = 300, scope = 1, quotas, scale = "scope",
  the_port = "ASTORIA / WARRENTON", prob_type = "type_clust_perc")
}
\arguments{
\item{input}{Overall input to the function; defaults to filt_clusts which is the data with clusters
filtered to include clusters with the most tows}

\item{ntows}{Number of tows in the fishing trip}

\item{start_clust}{Starting cluster to fish in}

\item{seed}{Seed for random sampling}

\item{scope}{Scope of available information; probably will be taken out}

\item{quotas}{Data frame of quota for target and weak stock species}

\item{scale}{Scale of movement; if scale == 'port', specify a d_port; if scale == 'scope', 
specify scale}

\item{prob_type}{Probability type; type_prop_hauls -- frequency of encounter 
or type_clust_perc -- proportion of catch in each cluster}
}
\description{
Function to conduct a fishing trip
Function to conduct a fishing trip
}
