% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_trip.R
\name{one_trip}
\alias{one_trip}
\title{Sample hauls for one trip
Function to sample hauls for one trip}
\usage{
one_trip(filt_data, cluster_value, num_hauls, unq_bin_scope = 0, quotas)
}
\arguments{
\item{filt_data}{Filtered data input into function; This data should not be vessel-specific}

\item{cluster_value}{Cluster specified to fish in}

\item{num_hauls}{Number of hauls to sample in each cluster (or rather each clust_bin)}

\item{unq_bin_scope}{Scope of bins to sample}

\item{quotas}{Quotas to  track}
}
\description{
Sample hauls for one trip
Function to sample hauls for one trip
}
