% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dummys.R
\name{process_dummys}
\alias{process_dummys}
\title{Process dummy variables
Function to calculate revenues and distances}
\usage{
process_dummys(xx)
}
\arguments{
\item{xx}{Index value}
}
\description{
Process dummy variables
Function to calculate revenues and distances
}
