% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dummys2.R
\name{process_dummys2}
\alias{process_dummys2}
\title{Process dummy variables
Function to calculate revenues and distances}
\usage{
process_dummys2(xx, td2 = td1, dat1 = dat, hab_dist = 5, n_cost)
}
\arguments{
\item{xx}{Index value}

\item{td2}{The tow dates}

\item{dat1}{The Data}

\item{hab_dist}{Distance used to filter for habits and revenues}

\item{net_cost}{Way of netting out costs for revenue calclations}
}
\description{
Process dummy variables
Function to calculate revenues and distances
}
