% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rum_data.R
\name{rum_data}
\alias{rum_data}
\title{Format Data for Random Utility Model
Function to format data for random utility model. Each row should be a cluster
and columns represent arguments for each coefficient}
\usage{
rum_data(the_input, the_quotas = quotas, risk_aversion)
}
\arguments{
\item{the_input}{Data for input}

\item{the_quotas}{Quota amounts, used to determine how close to quota a vessel is}

\item{risk_aversion}{The extent to which vessel avoids weak stock species}
}
\description{
Format Data for Random Utility Model
Function to format data for random utility model. Each row should be a cluster
and columns represent arguments for each coefficient
}
