% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_catch.R
\name{summarize_catch}
\alias{summarize_catch}
\title{Summarize catch
Function to take haul_id, and count the things that came out}
\usage{
summarize_catch(input = filt_clusts, haul_id)
}
\arguments{
\item{input}{Data as input; defaults to filt_clusts}

\item{haul_id}{Haul_id value}
}
\description{
Summarize catch
Function to take haul_id, and count the things that came out
}
