% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when_spp_delta.R
\name{when_spp_delta}
\alias{when_spp_delta}
\title{Function to calculate values for delta plots
Skew and proportion of zeroes specifically}
\usage{
when_spp_delta(data_type = tows_clust, period, spp)
}
\arguments{
\item{data_type}{Type of data; either tows_clust, or logbook}

\item{when}{Time period to focus on}

\item{skew}{Skew of non-zero, logged pound values}
}
\description{
Function to calculate values for delta plots
Skew and proportion of zeroes specifically
}
